refactor(k8s): split kubeconfig parsing by concern
