+++ Packages/K8sClient/Sources/K8sClient/KubeConfig/KubeConfigError.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// Errors from kubeconfig parsing.
+public enum KubeConfigError: Error, Sendable {
+    case missingField(String)
+    case invalidCertificate(String)
+    case execPluginFailed(String)
+}
